Move Ref and subclasses to to their own files
