+++ tests/run-with-compiler-custom-args/tasty-interpreter/interpreter/abstr/Eager.scala
@@ -1,0 +1,3 @@
+package scala.tasty.interpreter.abstr
+
+class Eager(val get: Any) extends Ref
