+++ tests/run-with-compiler-custom-args/tasty-interpreter/interpreter/abstr/Lazy.scala
@@ -1,0 +1,5 @@
+package scala.tasty.interpreter.abstr
+
+class Lazy(thunk: => Any) extends Ref {
+  lazy val get: Any = thunk
+}
