+++ tests/run-with-compiler-custom-args/tasty-interpreter/interpreter/abstr/Ref.scala
@@ -1,0 +1,5 @@
+package scala.tasty.interpreter.abstr
+
+trait Ref {
+  def get: Any
+}
