+++ tests/run-with-compiler-custom-args/tasty-interpreter/interpreter/abstr/Var.scala
@@ -1,0 +1,9 @@
+package scala.tasty.interpreter.abstr
+
+class Var(private var value: Any) extends Ref {
+  def get = value
+
+  def update(rhs: Any): Unit = {
+    value = rhs
+  }
+}
